Split focus/ranking.rs into focused sub-modules (#564)
Decompose the ~1,250-line ranking.rs into a ranking/ directory with
four focused sub-modules, each well under 1,500 lines:

- ranking/mod.rs (519 lines) — public API, orchestration, rank_focus_neurons
- ranking/record_providers.rs (161 lines) — RecordProvider trait, eager/lazy impls
- ranking/score_calculation.rs (200 lines) — neuron score computation, frequency
- ranking/removal_candidates.rs (315 lines) — removal candidates, SynapseCounts

All existing tests pass without modification. No public API changes.
